[#38] feat: index, chatting 화면 분리
